Add optional editor argument
I want to be able to set a non default editor in pgcli rather than use the default
from the eviroment variables. Luckily the Click library provides for a non default
editor in it's click.edit function. I am therefore making the change here so that
I can later make the required change in the pgcli project.