Fix test_slash_l* to support non-en_US locales
Update the tests to expect the current locale rather than `en_US`.
While technically this isn't 100% guaranteed to be the same locale
as the database is running on, I think we can reasonably assume that
for the purpose of testing they are the same.  Non-UTF-8 locales
still are not supported but the other tests seem to be unable to handle
them well anyway.

Fixes #140